requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.3
gensim==3.8.3
idna==2.10
joblib==0.16.0
murmurhash==1.0.2
numpy==1.19.2
plac==1.1.3
preshed==3.0.2
python-crfsuite==0.9.7
pyvi==0.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn-crfsuite==0.3.6
smart-open==2.2.0
spacy==2.3.2
srsly==1.0.2
tabulate==0.8.7
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.49.0
urllib3==1.25.10
vi-spacy-model==0.2.1
wasabi==0.8.0
